clean up output format for automated plotting

+++ basic_network_measures/ip.c
@@ -368,7 +368,7 @@ static void print_report( void )
     double cycles_per_sec = get_cpu_mhz(0) * 1000000;
     double cpu_usage = (ccompleted - cposted) / (cycles_per_sec * (elapsed / 0x1000000) );
 
-    printf("cycles used: %lu\n", ccompleted - cposted);
+    printf("cycles used: %llu\n", (unsigned long long) ccompleted - cposted);
     printf("cycles per second %f\n", cycles_per_sec);
     printf("time elapsed %2.7f sec\n", elapsed/ 0x1000000);
     printf("all cycles in time elapsed %f\n", cycles_per_sec * (elapsed / 0x1000000));
